galapagos gui: Begin major redesign of first test gui: dedicated editors for pattern, kernel, packages, and test procedures. Under construction, temporarily disabled linkage to driver library.
git-svn-id: https://subversion.gsi.de/dabc/drivers@4413 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ galapagos/gui/qt/framework/BasicObject.h
@@ -16,7 +16,7 @@ class BasicObject
   /** user readable name, same as in sequencer control window */
   std::string fName;
 
-  BasicObject(uint16_t id, const char* name): fId(id), fName(name){}
+  BasicObject(uint32_t id, const char* name): fId(id), fName(name){}
 
   BasicObject(const BasicObject& ob)
   {
